requirements.txt
astra-engine>=0.0.2b6
python-dotenv>=1.0.0
google-genai>=0.2.2
aiohttp>=3.9.0
motor>=3.3.2
aiosqlite>=0.19.0
psutil>=5.9.0
yt-dlp>=2023.12.30
sphinx>=7.0.0
sphinx-rtd-theme>=2.0.0
flask>=2.3.0
flask-restful>=0.3.0
requests>=2.31.0
